refactor: split useDomainTokenBalance up to sit in correct locations
